refactor(#36): OptimisticLock を Models/Traits に移動し Observer を削除
- OptimisticLock trait を app/Support/Models/OptimisticLock/ から app/Support/Models/Traits/ へ移動
- namespace を App\Support\Models\Traits に更新
- Observer 廃止に合わせ exclusion() をトレイトのメソッドに集約
- OptimisticLockObserver を削除
