+++ backends/php-laravel/app/Support/Models/AppModel.php
@@ -14,7 +14,7 @@ use DateTimeInterface;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use App\Support\Models\OptimisticLock\OptimisticLock;
+use App\Support\Models\Traits\OptimisticLock;
 
 /**
  * 基底Modelクラスです。
