--- backends/php-laravel/app/Support/Models/OptimisticLock/OptimisticLockObserver.php
@@ -1,77 +1,0 @@
-<?php
-
-/**
- * This is a program developed by Strategic Insights, Inc.
- *
- * Copyright (c) Strategic Insights, Inc. All Rights Reserved.
- */
-
-declare(strict_types=1);
-
-namespace App\Support\Models\OptimisticLock;
-
-use App\Support\Exceptions\AppException;
-use App\Support\Models\AppModel;
-
-/**
- * 楽観的ロックObserverクラスです。
- *
- * @author Satoshi Nagashiba <nagashibas@sii-japan.co.jp>
- * @package App\Support\Models\OptimisticLock
- */
-class OptimisticLockObserver
-{
-    /**
-     * 登録前の排他制御を行います。
-     *
-     * @param AppModel $model モデル: void
-     * @return void
-     */
-    public function created(AppModel $model): void
-    {
-        //
-    }
-
-    /**
-     * 更新前の排他制御を行います。
-     *
-     * @param AppModel $model モデル
-     * @return void
-     */
-    public function updating(AppModel $model): void
-    {
-        $this->exclusion($model);
-        $model->version++;
-    }
-
-    /**
-     * 論理削除前の排他制御を行います。
-     *
-     * @param AppModel $model モデル
-     * @return void
-     */
-    public function deleting(AppModel $model): void
-    {
-        $this->exclusion($model);
-        $model->version++;
-    }
-
-    /**
-     * 排他を確認します。
-     *
-     * @param AppModel $model モデル
-     * @return void
-     */
-    private function exclusion(AppModel $model): void
-    {
-        //現在バージョンを取得します
-        $key = $model->getKeyName();
-        $current = $model->withTrashed()->find($model->$key);
-
-        //更新されているかチェック
-        if ($model->getOriginal('version') != $current?->version) {
-            //楽観ロック
-            throw AppException::internal('optimistic lock');
-        }
-    }
-}
